third_party/blink/web_tests/fast/multicol/vertical-lr/balancing/balance-trailing-border-after-break-expected.txt

Tests that the bottom border at a column boundary in a block that lives in multiple columns is recorded properly for balancing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mc.offsetWidth is 69
PASS successfullyParsed is true

TEST COMPLETE


There should be no space between the short black line and the long gray one.
